Refactor SOS code, Added setup for player and platform implementation, no collision yet. Also updated SDL interfacing, removed Graphics class for SDL, since it works different. PL(PetaLinux) will have its own class initialized in main.cpp

+++ SOS/include/game.h
@@ -7,12 +7,15 @@
 #include "Object.h"
 #include <iostream>
 
+#include "objects/player.h"
+#include "objects/platform.h"
+
 class Game {
 public:
     Game();
     ~Game();
 
-    void update(float deltaTime);
+    void update(uint64_t deltaTime);
     void render();
     bool isRunning() const;
     std::vector<Object*>& getObjects();
